Scraper last track is added only once
Scrapper make hits every minute due to that it is possible to read the same song for few times. Now the song will be written only once to the db when countinously reading the same song. When the song will be played in the future it will be added again to the sub db.